se termina de agregar el token a todas las apis

+++ src/main/java/org/sofka/demo/api/controller/CountryController.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 
 import org.sofka.demo.domain.Country;
 import org.sofka.demo.repository.CountryRepository;
+import org.sofka.demo.repository.UserRepositoryIpm;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,16 +19,22 @@ import org.springframework.web.bind.annotation.*;
 public class CountryController {
 	@Autowired
 	private CountryRepository countryRepository;
-	
+
+	@Autowired
+	private UserRepositoryIpm userRepositoryIpm;
+
 	@GetMapping("/countries")
-	public List<Country> findAllCountries() {
+	public List<Country> findAllCountries(@RequestHeader(value="Authorization") String token) {
+		if (!userRepositoryIpm.validateToken(token)) { return null; }
 		List<Country> countries = new ArrayList<>();
 		countryRepository.findAll().forEach(countries::add);
 		return countries;
 	}
 	
 	@GetMapping("/country/{code}")
-	public ResponseEntity<Country> findCountryByCode(@PathVariable(name = "code") String code) {
+	public ResponseEntity<Country> findCountryByCode
+			(@PathVariable(name = "code") String code, @RequestHeader(value="Authorization") String token) {
+		if (!userRepositoryIpm.validateToken(token)) { return null; }
 		Optional<Country> country = countryRepository.findCountryByCode(code);
 		if (country.isPresent()) {
 			return ResponseEntity.ok().body(country.get());
@@ -37,12 +44,14 @@ public class CountryController {
 	}
 	
 	@PostMapping("/newCountry")
-	public Country saveCountry(@Validated @RequestBody Country country) {
+	public Country saveCountry(@Validated @RequestBody Country country, @RequestHeader(value="Authorization") String token) {
+		if (!userRepositoryIpm.validateToken(token)) { return null; }
 		return countryRepository.save(country);
 	}
 	
 	@PutMapping("/country/{id}")
-	public Country updateCountry(@RequestBody Country newCountry, @PathVariable int id) {
+	public Country updateCountry(@RequestBody Country newCountry, @PathVariable int id, @RequestHeader(value="Authorization") String token) {
+		if (!userRepositoryIpm.validateToken(token)) { return null; }
 		return countryRepository.findById(id).map(country -> {
 			country.setName(newCountry.getName());
 			country.setCode(newCountry.getCode());
@@ -54,7 +63,8 @@ public class CountryController {
 	}
 	
 	@DeleteMapping("/api/country/{id}")
-	public ResponseEntity<Void> deleteCountry(@PathVariable int id) {
+	public ResponseEntity<Void> deleteCountry(@PathVariable int id, @RequestHeader(value="Authorization") String token) {
+		if (!userRepositoryIpm.validateToken(token)) { return null; }
 		if (countryRepository.existsById(id))
 			countryRepository.deleteById(id);
 		else
